Introduce customized gever table sources/configs.
That way we can add functionality to tabbed-view/table
sources and configs in one way. This should also
streamline the inheritance hierarchy used for new
tabs a bit.

Also move support for filtering the list to base table
source implementation.
