+++ opengever/tabbedview/interfaces.py
@@ -1,3 +1,5 @@
+from ftw.table.interfaces import ICatalogTableSourceConfig
+from ftw.table.interfaces import ITableSourceConfig
 from zope.interface import Interface
 
 
@@ -5,5 +7,9 @@ class ITabbedViewEnabled(Interface):
     """TabbedView behaviour"""
 
 
-class IFilterListTableSourceConfig(Interface):
+class IGeverTableSourceConfig(ITableSourceConfig):
+    pass
+
+
+class IGeverCatalogTableSourceConfig(IGeverTableSourceConfig, ICatalogTableSourceConfig):
     pass
