+++ opengever/tabbedview/sqlsource.py
@@ -1,39 +1,25 @@
 from five import grok
-from opengever.tabbedview import GeverTableSource
+from ftw.table.basesource import BaseTableSource
+from ftw.table.interfaces import ITableSource
+from opengever.tabbedview.filtered_source import FilteredTableSourceMixin
+from opengever.tabbedview.interfaces import IGeverTableSourceConfig
 from sqlalchemy import or_
 from sqlalchemy.orm.query import Query
 from sqlalchemy.sql.expression import asc
 from sqlalchemy.sql.expression import column
 from sqlalchemy.sql.expression import desc
+from zope.interface import Interface
 
 
-class SqlTableSource(GeverTableSource):
+class SqlTableSource(grok.MultiAdapter, FilteredTableSourceMixin, BaseTableSource):
     """Base table source adapter for every listing,
        that gets the content from sql.
     """
 
-    grok.baseclass()
-
     searchable_columns = []
 
-    def build_query(self):
-        """Builds the query based on `get_base_query()` method of config.
-        Returns the query object.
-        """
-        # initalize config
-        query = super(SqlTableSource, self).build_query()
-
-        # reviewstate-filter
-        if self.config.filterlist_available:
-            query = self.extend_query_with_filter(query)
-
-        return query
-
-    def extend_query_with_filter(self, query):
-        """When the filterlist is active, we update the query with
-        the current filter."""
-        selected_filter_id = self.request.get(self.config.filterlist_name)
-        return self.config.filterlist.update_query(query, selected_filter_id)
+    grok.implements(ITableSource)
+    grok.adapts(IGeverTableSourceConfig, Interface)
 
     def validate_base_query(self, query):
         """Validates and fixes the base query. Returns the query object.
